Fix up controller tests.

+++ src/Auth/AclAdapter/DbAclAdapter.php
@@ -67,7 +67,7 @@ class DbAclAdapter implements AclAdapterInterface {
 			list($plugin, $key) = explode('.', $key, 2);
 		}
 		if (strpos($key, '/') !== false) {
-			$pos = strrpos($key, '/');
+			$pos = (int)strrpos($key, '/');
 			$prefix = substr($key, 0, $pos);
 			$key = substr($key, $pos + 1);
 		}
